refactor : fetchWithTokenRenewal 함수 구현 완료
